Added additional widgets to the status bar.
This commit adds additional labels to the status bar which display the
current line and column number of the cursor, as well as a widget which
will be used to display notifications.

Additionally, this patch adds borders around the various status bar
fields, to visually separate them from each other.